# public/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "ProjectX Gateway API",
    "description": "ProjectX Gateway API Documentation and specification.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "produces": [
    "text/plain",
    "application/json",
    "text/json"
  ],
  "paths": {
...
    "/api/Auth/loginApp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login as the specified user using the specified application.",
        "operationId": "Auth_LoginApp",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "The login request.",
            "schema": {
              "$ref": "#/definitions/LoginAppRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          }
        }
      }
    },
    "/api/Auth/loginKey": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login as the specified user using the specified API key.",
        "operationId": "Auth_LoginKey",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "The login request.",
            "schema": {
              "$ref": "#/definitions/LoginApiKeyRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          }
        }
      }
    },
...
    "/api/Contract/search": {
      "post": {
        "tags": [
          "Contract"
        ],
        "summary": "Searches for contracts based on the provided search criteria.",
        "operationId": "Contract_SearchContracts",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "The search criteria for finding contracts.",
            "schema": {
              "$ref": "#/definitions/SearchContractRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response containing the search results, including any matching contracts.",
            "schema": {
              "$ref": "#/definitions/SearchContractResponse"
            }
          }
        }
      }
    },
    "/api/Contract/searchById": {
      "post": {
        "tags": [
          "Contract"
        ],
        "summary": "Searches for a contract by its ID.",
        "operationId": "Contract_SearchContractById",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "The request containing the contract ID.",
            "schema": {
              "$ref": "#/definitions/SearchContractByIdRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response containing the search result, including the matching contract if found.",
            "schema": {
              "$ref": "#/definitions/SearchContractByIdResponse"
            }
          }
        }
      }
    },
    "/api/Contract/available": {
      "post": {
        "tags": [
          "Contract"
        ],
        "summary": "Lists available contracts based on the provided request parameters.",
        "operationId": "Contract_AvailableContracts",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "The listing criteria for available contracts.",
            "schema": {
              "$ref": "#/definitions/ListAvailableContractRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response containing the list of available contracts.",
            "schema": {
              "$ref": "#/definitions/ListAvailableContractResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "LoginResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/LoginErrorCode"
        },
        "errorMessage": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "LoginErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = UserNotFound\n2 = PasswordVerificationFailed\n3 = InvalidCredentials\n4 = AppNotFound\n5 = AppVerificationFailed\n6 = InvalidDevice\n7 = AgreementsNotSigned\n8 = UnknownError\n9 = ApiSubscriptionNotFound\n10 = ApiKeyAuthenticationDisabled",
      "x-enumNames": [
        "Success",
        "UserNotFound",
        "PasswordVerificationFailed",
        "InvalidCredentials",
        "AppNotFound",
        "AppVerificationFailed",
        "InvalidDevice",
        "AgreementsNotSigned",
        "UnknownError",
        "ApiSubscriptionNotFound",
        "ApiKeyAuthenticationDisabled"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6,
        7,
        8,
        9,
        10
      ]
    },
    "LoginAppRequest": {
      "type": "object",
      "required": [
        "userName",
        "password",
        "deviceId",
        "appId",
        "verifyKey"
      ],
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "deviceId": {
          "type": "string"
        },
        "appId": {
          "type": "string"
        },
        "verifyKey": {
          "type": "string"
        }
      }
    },
    "LoginApiKeyRequest": {
      "type": "object",
      "required": [
        "userName",
        "apiKey"
      ],
      "properties": {
        "userName": {
          "type": "string"
        },
        "apiKey": {
          "type": "string"
        }
      }
    },
...
    "SearchContractResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/SearchContractErrorCode"
        },
        "errorMessage": {
          "type": "string"
        },
        "contracts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContractModel"
          }
        }
      }
    },
    "SearchContractErrorCode": {
      "type": "integer",
      "description": "0 = Success",
      "x-enumNames": [
        "Success"
      ],
      "enum": [
        0
      ]
    },
    "ContractModel": {
      "type": "object",
      "required": [
        "id",
        "name",
        "description",
        "tickSize",
        "tickValue",
        "activeContract",
        "symbolId"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "tickSize": {
          "type": "number",
          "format": "decimal"
        },
        "tickValue": {
          "type": "number",
          "format": "decimal"
        },
        "activeContract": {
          "type": "boolean"
        },
        "symbolId": {
          "type": "string"
        }
      }
    },
    "SearchContractRequest": {
      "type": "object",
      "required": [
        "live"
      ],
      "properties": {
        "searchText": {
          "type": "string"
        },
        "live": {
          "type": "boolean"
        }
      }
    },
    "SearchContractByIdResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/SearchContractByIdErrorCode"
        },
        "errorMessage": {
          "type": "string"
        },
        "contract": {
          "$ref": "#/definitions/ContractModel"
        }
      }
    },
    "SearchContractByIdErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = ContractNotFound",
      "x-enumNames": [
        "Success",
        "ContractNotFound"
      ],
      "enum": [
        0,
        1
      ]
    },
    "SearchContractByIdRequest": {
      "type": "object",
      "required": [
        "contractId"
      ],
      "properties": {
        "contractId": {
          "type": "string"
        }
      }
    },
    "ListAvailableContractResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/ListAvailableContractErrorCode"
        },
        "errorMessage": {
          "type": "string"
        },
        "contracts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContractModel"
          }
        }
      }
    },
    "ListAvailableContractErrorCode": {
      "type": "integer",
      "description": "0 = Success",
      "x-enumNames": [
        "Success"
      ],
      "enum": [
        0
      ]
    },
    "ListAvailableContractRequest": {
      "type": "object",
      "required": [
        "live"
      ],
      "properties": {
        "live": {
          "type": "boolean"
        }
      }
    },
...
  }
}